Composer 미디어 lifecycle 책임을 분리한다
